Fix SMALLINT overflow on episode_number/season_number
Some long-running daily series exceed SMALLINT's 32767 cap (e.g.
episode_number 48501). Widen both columns to INTEGER, mirroring the
earlier runtime_minutes fix.

+++ etl/main.py
@@ -160,8 +160,8 @@ def create_staging_tables(conn):
             CREATE TABLE episodes_new (
                 tconst VARCHAR(10) NOT NULL,
                 parent_tconst VARCHAR(10) NOT NULL,
-                season_number SMALLINT,
-                episode_number SMALLINT
+                season_number INTEGER,
+                episode_number INTEGER
             )
         """)
         cur.execute("""
